Address Copilot review on PR #248: extract RecordingScanner to neutral module
Move RecordingScanner from tests/controllers/devices/conftest.py to
tests/_recording_scanner.py (sibling of _storage_fixtures.py). Both
the devices conftest and the firmware test_refresh.py now import
from the neutral module, dropping the cross-conftest dependency
that Copilot flagged.
